Updated openstack/openstack Project: openstack/nova  594550982efc5d49378ca90a7292df86e6508f88
Remove usage of remotable context parameter in service, tag, vif

This removes the use of the context parameter in remotable methods
in these objects. This is a precursor to actually removing the
context from the call, but we need to remove dependence on the
argument first. Remotable methods should be using the bundled object
context.

Related to blueprint kilo-objects

Change-Id: I8d040a03d0daa314e8fc47747ba8c7735fd744d0