Updated openstack/openstack Project: openstack/nova  4ec517e2966a90d9661f48e9f28d04cff71b893f
Actually stop passing context to remotable methods

This actually passes None for the context parameter to remotable methods,
ensuring that they cannot regress dependence on it. It also fixes up the
base object tests to be in line with this, and removes the test that verifies
that alternate contexts can be passed.

Related to blueprint kilo-objects

Change-Id: I5a7ab4b157e8b6217644934f60ff17222e6200b9