Updated openstack/openstack Project: openstack/cinder  a0f8cc6bca87f99c4a64e7e967226bdfd8877ad1
Sort snapshots in create CG from CG snapshot

This problem was discovered while testing create CG from CG snapshot.
In create_consistencygroup_from_src() in manager.py, it calls the driver
method create_consistencygroup_from_src() and passes in parameters
including source snapshots and target volumes. The problem is that the
snapshots list is not sorted to be in the same order as their corresponding
volumes. So they may be in the wrong order.

In this patch, the source snapshots were sorted before being passed to
the driver.

Change-Id: I988ce2836f2d553391deeee6e68738356877109b
Closes-Bug: #1430628