Updated openstack/openstack Project: openstack/horizon  2e4d2a0af197c55062a2853ab7fae40e1a15ff7f
Source step - fix donut chart label color

On the source step of the new Launch Instance Wizard, the label
for the donut chart does not match the color in the mockups.
The current code uses the "color" attribute to set the label color,
but it actually should be the "fill" attribute, as that is what is
used by SVG to render the label.

Closes-Bug: 1433669

Change-Id: Iadfda86007932b48191e0abfa846646cc66e4213