Updated openstack/openstack Project: openstack/python-keystoneclient  bde71a25f11a0ad067df6c648c3b550254c825d5
Allow passing logger object to request

It can become difficult to trace the client that created HTTP requests
as the logging all goes through the keystoneclient.session logger. Allow
passing the logger through the request function and make it able to be
set via the adapter so it can be set once per client instantiation.

Change-Id: Id45c315bee9a56f1c241210d667470751bf689d5
Closes-Bug: #1421868