Updated openstack/openstack Project: openstack/nova  75166ea49d2de9664c9c7ef14b8fb7fb0ac1799a
Expand help message on some quota config options

There are two config options related to keeping quotas in sync which
could be very useful for deployers who are having issues with that.
They're quietly tucked away at the moment and it may not be clear that
they should be used.  This expands the help message to explain why they
default to off and that it may be useful to turn them on.

Change-Id: I27a08bacf01aa2a4b285b50df97337c985291195