Updated openstack/openstack Project: openstack/horizon  4ced35db76507a9804cbf180eec094a6900626e7
Improve styling for key pair detail table

In the Access and Security step of the new Launch Instance Wizard,
there is a detail table that shows when expanding the key pairs.
Currently the key pair public key is displayed within a textarea with
a white background. This patch removes the textarea and outputs the
public key into pre and code tags so that it follows standard design
patterns/conventions, and also is responsive in width and bkg color

Closes-Bug: 1433696
Change-Id: I6d1cd7892ae8301c879e874ddabf31cc22fb8f24