Updated openstack/openstack Project: openstack/heat  7e67aff769e5cdafc4587d2c5291c627aff958f1
Fix config drive not created when boot from volume

When booting a instance from volume, destination_type is needed
to set to volume in block device mapping v2. However, parameter
block device mapping v2 is not in heat when booting an instance,
so we need to add it as property for OS::Nova::Server.

Change-Id: Ie56358ed6fdb7cda4a28d546817c31e71ca179b3
Closes-Bug: #1358624