Updated openstack/openstack Project: openstack/nova  2413aa8a857c6e1512054f5adabec55599665054
Add InstanceMapping object

This adds the InstanceMapping object for interacting with the
instance_mapping table.

Rather than creating a new db api the database operations are embedded
within the object.  This keeps the logic near where it's being used.
This does make an assumption about using sqlalchemy for now.  In order
to allow for substituting the database driver we will probably want to
setup a subclassing system for objects to swap that part out.  This
would allow for finer control than we currently have.

Change-Id: Ide9cc7a255b886647fd424c2030beff0335e2cd3
bp: cells-v2-mapping