Updated openstack/openstack Project: openstack/nova  b9beb69b27cc6f51573f3891dbc6f3904554e36b
Add CellMapping object

This introduces the first object for the cellsv2 api database.  Rather
than creating a new db api the database operations are embedded within
the object.  This keeps the logic near where it's being used.

This does make an assumption about using sqlalchemy for now.  In order
to allow for substituting the database driver we will probably want to
setup a subclassing system for objects to swap that part out.  This
would allow for finer control than we currently have.

Change-Id: Ie4828ef02e23047c4ec88e269e3a9b21b113b9e5
Partially-Implements: bp cells-v2-mapping