Updated openstack/openstack Project: openstack/tripleo-heat-templates  558cd7d09ce0604dd744a3fd2166ce361ad90cc2
puppet: tidy up the compute nova neutron config

Updates the puppet configuration for the Nova neutron configs
so that we do the join conversions in the Heat templates and
use only hiera for configuration of the ::nova::network::neutron
class. This updates the compute configuration to match what
we now do on the controller as well.

Change-Id: I2b352551777f64e0ceb119f48cc3b3ab1779f4d5