Updated openstack/openstack Project: openstack/nova  0ff58bbea89b7643edc54c3bb36d17a61b825ade
Fix up join() and leave() methods of servicegroup

The join() method was not documented properly, and the Zookeeper
driver's implementation of the join() method returned a FakeLoopingCall
object for no reason whatsoever (the other drivers just return None), so
it's not necessary.

The leave() method wasn't used anywhere at all, therefore this patch
removes it entirely.

Change-Id: Idbd95636dbebfb4f21a25d23625ee338c1ba863e
Closes-bug: #1414513
Closes-bug: #1414515