Updated openstack/openstack Project: openstack/neutron  d0589051adec39fc3d70554aa919db0bedfd73b2
Add a netns-cleanup functional test

We have lots of regressions on netns, because unit test
is not enough. This commit adds basic functional testing
to the netns_cleanup.

This work should be extended when we have functional testing
for the dhcp agent, spawning dhcp services, and then
making sure they're fully cleaned up.

Change-Id: I0b5125dfa24a3dbcd44593ae2aee4dbbd47def67