Updated openstack/openstack Project: openstack/nova  db59ba3a22a49aec6d9c1d657193ed07f39e261a
libvirt: partial fix for live-migration with config drive

In current version of nova, live-migration with config drive on
local disk is forbidden due to the bug of libvirt of copying readonly
disk. However, if we use vfat as the format of config drive,
the function of live-migration works well.

In this patch, we re-open the function of live-migration to admin users.
Notice you should add 'config_drive_format=vfat' in nova.conf explicitly.

This patch doesn't solve the problem fundamentally which need further
efforts, but offers a simple and feasible workaround to user.

DocImpact

Co-Authored-By: Davanum Srinivas <davanum@gmail.com>
Change-Id: I7429e12766da7f7f8d484b3a9df6247e832816b0
Partial-Bug: #1246201