Updated openstack/openstack Project: openstack-infra/project-config  a175e5ad332aa021e07855714b88553d14b25125
Fix failure detection in irc access check

Two problems are corrected:

1) If a channel is not registered with chanserv, consider it an
access failure.

2) The failure flag was being reset on every successful private
message.  That was intended to ensure that we failed if we were
unable to communicate with ChanServ.  But the logic was flawed.
Correct this by treating the failure flag as trinary.  On the
first channel where we have verified access is okay, set the flag
to indicate that we have not failed, iff the flag has not already
been set.  So now either failure to communicate with ChanServ or
an actual access failure will cause it to exit 1.

Change-Id: I8d51b706ed6f499827f8046ab4522b342589cb90