Updated openstack/openstack Project: openstack/nova  c71b4749f28768310202e08f59115a7d3030e7cd
Deprecate Nova in tree EC2 APIs

The EC2 APIs have been poorly maintained for multiple cycles. The
focus and effort has shifted to the stackforge ec2-api project
http://git.openstack.org/cgit/stackforge/ec2-api/. We have landed Nova
API changes required to support this in Kilo.

This will give significant warning for people during the Kilo release
so we can address any possible issues with the ec2-api external
support. It also reflects the reality that these interfaces are not
well maintained in Nova, and hopefully resets expectations correctly
on them.

UpgradeImpact

DocImpact

Change-Id: I6726454107f9f8ad5eaaf7a5dc0e8193ce73bd66