Updated openstack/openstack Project: openstack/horizon  0e7a809d9431cbb357e1e4044bef23f695235b2a
Allow specifying authentication URLs from settings

To allow us to have authentication mechanisms that live out of tree as
AUTHENTICATION_BACKENDS do we need to have a way to load the views that
service those backends.

Provide a means via settings that we can add new routes to the /auth
urls.

Closes-Bug: #1433358
Change-Id: I1ffc8a4e0bbfa58c7b50351432668d49d3a7dab1