Updated openstack/openstack Project: openstack/cinder  6b12a4281c1f1a89806908571a3b1af35cd17520
Windows iSCSI driver: Fixes copy disk method exception handling

In the copy_vhd_disk method, it is assumed that if the copy operation
fails, a wmi.x_wmi exception will be raised. This is wrong. In fact,
none of the methods called there can raise such an exception in those
circumstances.

At the moment, this method will not raise an exception even if the
copy operation fails.

This patch adds a check to see if the source image actually exists.
The according CIM_DataFile Copy method's return value is checked as
well in order to see if the operation failed or not.

Change-Id: I778ecd61adb007995b629b36044caac78cd16889
Closes-Bug: #1356398