Updated openstack/openstack Project: openstack/neutron  0be020d3ddb568147df1518d3b47d7fbd2e6e2bf
Include IPv6 SLAAC addresses implicitly for port update

(Patch set #4 for the multiple-ipv6-prefixes blueprint)

This change adds functional code and UT for port-update
handling changes as described in the Neutron blueprint
'multiple-ipv6-prefixes'.

Whenever a port is updated on a network that includes one
or more auto-address subnets (e.g. SLAAC or DHCPv6-stateless),
then any addresses that the port previously had for those
subnets needs to be implicitly retained for the update
operation. This patch set adds this implicit retention
of the auto-addresses for port update operation.

Change-Id: I91687d05a48bc3938e4bf34a7c454823a500730c
Partially-implements: blueprint multiple-ipv6-prefixes