Updated openstack/openstack Project: openstack/nova  498518f8562a17e78ffa7efbb09c96add8f6776a
BDM: Avoiding saving if there were no changes

Since objects currently only check if the attribute was set not that the
actual value was changed (and it's not a straightforward task in the
general case), we can still make this better in the DriverBlockDevice
classes and avoid unnecessary DB traffic while being able to make our
saves on as granular level as possible.

This patch makes the code check for changes before actually setting them
on the underlying object.

Change-Id: Ib00ab1a275baabf7d2ea7ad988183b4727905a70