Updated openstack/openstack Project: openstack/nova  35c05a4070f43f2bf0fe876bf1a0813e6a764cb8
Save bdm.connection_info before calling volume_api.attach_volume

There is a race in attach/detach of a volume where the volume status
goes to 'in-use' before the bdm.connection_info data is stored in the
database. Since attach is a cast, the caller can see the volume go to
'in-use' and immediately try to detach the volume and blow up in the
compute manager because bdm.connection_info isn't set stored in the
database.

This fixes the issue by saving the connection_info immediately before
calling volume_api.attach_volume (which sets the volume status to
'in-use').

Closes-Bug: #1327218

Change-Id: Ib95c8f7b66aca0c4ac7b92d140cbeb5e85c2717f