Updated openstack/openstack Project: openstack/horizon  755eaf105a030a46602d6d523eeaed0a1d821f93
Improve styling for security group detail table

In the Access and Security step of the new Launch Instance Wizard,
there is a detail table that shows when expanding the security groups.
Currently this detail table is very compact/congested and uses table
striping, which does not match the mockups. This patch removes table
striping from the security group details table, and adds padding so
that it more closely matches the mockups.

Closes-Bug: 1433688
Change-Id: Iccd4a83cb3c7e36383f6063cb65d9e7c4efe84b1