Updated openstack/openstack Project: openstack/cinder  f989577bd36a88c740ea99e04530cb3c12ca0361
Removing HP MSA driver for no reported CI

CI deadlines were set and pushed since last year. An email about this
requirement and the deadline of March 19th 2015 has been sent to each
individual driver maintainer, as well as the mailing list [1]. This
driver is being removed because it's not supported, so it won't have
a CI reporting to ensure their driver integration is successful.
Therfore, we can not validate the driver is working in Cinder today in
a continuous way.

DocImpact
[1]- http://lists.openstack.org/pipermail/openstack-dev/2015-January/054614.html

Change-Id: Ia6681e943ffb807ff6b3064976f6ef8eb3259e65