Updated openstack/openstack Project: openstack-dev/devstack  6f3f310848d1134ff73dd23e246ad00f7cd13365
Fix packages not getting installed if service name in base path

Currently, if devstack base path includes the name of a given
service (e.g. nova), then the service's prereq packages will
not be installed. This fix changes the checking the match
against the full path of the package list file rather than the
name of a given service.

Closes-Bug: #1434314
Change-Id: Ie81352ebd5691afc6d0019f71d5b62370e8bb95f