Updated openstack/openstack Project: openstack/cinder  ba379ee9d1fc7c4e75344d0cf050e412c6d67c7d
Removing ZFSSA driver for no reported CI

CI deadlines were set and pushed since last year. An email about this
requirement and the deadline of March 19th 2015 has been sent to each
individual driver maintainer, as well as the mailing list [1]. This
driver is being removed because the maintainer does not have a CI
reporting to ensure their driver integration is successful. Therfore, we
can not validate the driver is working in Cinder today in a continuous
way.

DocImpact
[1]- http://lists.openstack.org/pipermail/openstack-dev/2015-January/054614.html
Change-Id: I5126ee5c56d2a7382e67626d28197eafa5940be5