Updated openstack/openstack Project: openstack/cinder  a0cb4b97061d0cc8a709b83a5220ae7343c3aba4
Removing Symantec NFS driver for no reported CI

CI deadlines were set and pushed since last year. An email about this
requirement and the deadline of March 19th 2015 has been sent to each
individual driver maintainer, as well as the mailing list [1]. This
driver is being removed because the maintainer does not have a CI
reporting to ensure their driver integration is successful. Therfore, we
can not validate the driver is working in Cinder today in a continuous
way.

DocImpact
[1]- http://lists.openstack.org/pipermail/openstack-dev/2015-January/054614.html

Change-Id: I38b66f8df30508704df6081f18a09a3a12853858