Updated openstack/openstack Project: openstack/cinder  9c3f16c35a9bdcb24df6012554ef83c6da2f33c8
Removing FusionIO driver for no reported CI

CI deadlines were set and pushed since last year. An email about this
requirement and the deadline of March 19th 2015 has been sent to each
individual driver maintainer, as well as the mailing list [1]. This
driver is being removed because the maintainer has chosen not to respond
to the CI deadline email and report a CI to ensure their driver
integration is successful. Therfore, we can not validate the driver is
working in Cinder today in a continuous way.

DocImpact
[1] - http://lists.openstack.org/pipermail/openstack-dev/2015-January/054614.html

Change-Id: Ic5881eec84e3069f80543bb20b0132ddb20d90ad