Updated openstack/openstack Project: openstack/cinder  4b5e16e320562a69dd687a51b7f505a6e8897bce
Removing Fujitsu driver for no reported CI

CI deadlines were set and pushed since last year. An email about this
requirement and the deadline of March 19th 2015 has been sent to each
individual driver maintainer, as well as the mailing list [1]. This
driver is being removed because the maintainer has chosen not to respond
to the CI deadline email and report a CI to ensure their driver
integration is successful. Therfore, we can not validate the driver is
working in Cinder today in a continuous way.

DocImpact
[1] - http://lists.openstack.org/pipermail/openstack-dev/2015-January/054614.html

Change-Id: Ibe9a3cc9831a1ef4aa7cbdb15e09bced353c738a