Updated openstack/openstack Project: openstack/cinder  6cfc6c56f790db605e4471f5649bad754fc49a5a
Removing DRBD driver for no reported CI

CI deadlines were set and pushed since last year. An email about this
requirement and the deadline of March 19th 2015 has been sent to each
individual driver maintainer, as well as the mailing list [1]. This
driver is being removed because the maintainer does not have a CI
reporting to ensure their driver integration is successful. Therfore, we
can not validate the driver is working in Cinder today in a continuous
way.

DocImpact
[1]- http://lists.openstack.org/pipermail/openstack-dev/2015-January/054614.html

Change-Id: If4b119ec6ca80dcf28530e77e1f98bda04ba6eab