Updated openstack/openstack Project: openstack/cinder  0d906d72c8556b99301689ca75818fc58375668d
Removing Zadara driver for no reported CI

CI deadlines were set and pushed since last year. An email about this
requirement and the deadline of March 19th 2015 has been sent to each
individual driver maintainer, as well as the mailing list [1]. This
driver is being removed because the maintainer does not have a CI
reporting to ensure their driver integration is successful. Therfore, we
can not validate the driver is working in Cinder today in a continuous
way.

DocImpact
[1]-
http://lists.openstack.org/pipermail/openstack-dev/2015-January/054614.html

Change-Id: I016e22c857f520a7eb79e1ed5b36735a97f68b68