Updated openstack/openstack Project: openstack/cinder  6784848ae21f1aaf508ad246b7e9e2599d79881a
Removing Huawei drivers for no reported CI

CI deadlines were set and pushed since last year. An email about this
requirement and the deadline of March 19th 2015 has been sent to each
individual driver maintainer, as well as the mailing list [1]. This
driver is being removed because the maintainer does not have a CI
reporting to ensure their driver integration is successful. Therfore, we
can not validate the driver is working in Cinder today in a continuous
way.

DocImpact
[1]- http://lists.openstack.org/pipermail/openstack-dev/2015-January/054614.html

Change-Id: I0599a8721b14ba4f2a255188b71be6d62ae8cc1a