Updated openstack/openstack Project: openstack/nova  dfcfa82af96bae12265875cb313d58109e536480
VMware: use instance UUID as instance name

The patch set remove the instance_name parameter that was passed
to spawn and other methods. This parameter was used for the naming
of the rescue and resize VM's. The fact that those are now done on
the original instance enables us to drop this 'instance_name'
parameter.

This simplifies the code and unit tests.

Change-Id: I7262d0b69fc08144b5675cfeafb5a40668a472d0