Updated openstack/openstack Project: openstack/python-keystoneclient  3a76d5c211f72fd00450ec540e7e77e5d3e17771
Rename requests mock object in testing

It has been mentioned a number of times that the self.requests naming
for the requests_mock object is confusing between whether you are
actually sending a request or are mocking a request.

Rename all entries of the requests object to requests_mock.

This cleans up a couple of entries where the older register_uri format
was being used in favour of using the HTTP method as the requests_mock
method.

Change-Id: I315085b4088130b510f9dbd696011d983598372c