Updated openstack/openstack Project: openstack-infra/project-config  90a8e7e5baca4c65185089b1049371d431230dfa
Update puppet-setproxy to belong to Gozer group

Puppet-openstack is intended to be a place for reviewing
puppet modules to deploy OpenStack, and that's not the
case. It's a generic module to be set proxy services,
more related with downstream Infra usage.
Moving ownership to Gozer team to grant control of it.

Change-Id: I281ee9686f9ae874bf35989a52c6fc692c8a9e9b