Updated openstack/openstack Project: openstack/cinder  73f2c6f329ce1073a4ac27ce430d05ee05ac85fa
Fix SAN generic driver ssh whitespaced commands

SAN driver was sending commands over ssh with whitespaces between
every character resulting in commands like "u n a m e - s" instead
of "uname -s" (example from bug report).

As the bug reporter noted, the problem was that there were two joins
in the process of mounting the command. The first one got the
tuple and joined into a string. The second one joined a whitespace
between each character of the string.

Additionally, this behavior made utils.ssh_injection a bad check,
because it expected a list of args used in 'for arg in cmd_list' and
what it got was a single string. This resulted in the check being made
over the single characters instead of each arg.

All this is to explain why I chose to remove the first join instead of
the second one.

An unit test for this issue was created. SAN driver had no unit test.

Change-Id: I8e8624f2b82c26522936adf05e637a7bbe34a069
Closes-Bug: #1422604