Updated openstack/openstack Project: openstack/swift  9ae0a0702c4a32407b1d00e93bd1be01e2ec26bc
Fix a lack of method arguments at tempauth

This fixes a kind of trivial issue with TypeError caused by a lack
of method argument at calling.

When using 'Authorization' header with tempauth, tempauth calls
_get_user_groups with 3 arguments (includes self) on current code
but the method requires 4 arguments so it will go to fail with 500.

This patch fixes to take correct arguments and adds an unit test
for the case.

Closes-Bug:1434465

Change-Id: Ibf00c41fa8e5cfdaf49bc4d571e575d7f229ebfe