Updated openstack/openstack Project: openstack/sahara-image-elements  4cd8af1b4ef1fb9977d8977555ee44cce20798f4
Use the dib-utils repository too

Newer diskimage-builder versions require the external dib-utils
repository, as dib-run-parts has been moved there.

In preparation for using a newer diskimage-builder, checkout and use the
dib-utils repository; thus dib-run-parts from dib-utils will be used
instead of the one currently in diskimage-builder (if using a version
shipping it), which works the same way.
The dib-utils version used is the current last one, 0.0.9.

Change-Id: I7f537486279c40123b65d8206f9405774e9ebcbf