Updated openstack/openstack Project: openstack-infra/project-config  7086751c21c2074ded68f1a13cd60500021ff6c8
Change node param for ec2api rally job.

Zuul expects the job to run on "devstack" workers because of the
"dsvm" in the job name (there's a zuul parameter function which
matches those up for convenience), but in this job definition I
have specified "bare" workers so that's all Jenkins registered for
it. As a result, zuul doesn't see any of the worker types it expects
registering that job name. I must correct the "node" parameter for
the gate-rally-dsvm-ec2-api-fakevirt job definition in
openstack-infra/project-config:jenkins/jobs/ec2-api.yaml then it
should work.

Change-Id: If5740498cb059e8219b2b9db4a4c565d2bc1b04e