Updated openstack/openstack Project: openstack/requirements  737e0d8d3214347d01f51667d3b235e6e08f710e
Bump tempest-lib min version

tempest-lib 0.4.0 was recently released which includes the auth layer,
tokens clients, and credentials objects. This contains all the pieces
for consumers of tempest lib to write there own service clients which
projects will want to start using soon. However, this is required for
tempest to remove the equivalent pieces from in tree and use the lib
versions with: https://review.openstack.org/166022

Change-Id: I6ffc70364ca6e6f27e58b1ed3c371c51104307b7