Updated openstack/openstack Project: openstack/cinder  8d26a6bcfee44f0528ca36cc26a43e252aaafde7
update oslo policy to remove policy.d log spam

After oslo policy added config dirs it started spamming the logs
relentlessly on if you didn't have a policy.d directory on your
system. This is a very unfriendly upgrade to operators.

A fix was provided in the incubator just before incubator policy was
deleted. This syncs in that fix.

Change-Id: Idf9c63c52561acdfe5463fee35dbc497c5e1fde5
Related-Bug: #1421863