Updated openstack/openstack Project: openstack/horizon  ff565cd39a9d18e6cd851015227373ad2201fde2
Add placeholder text to transfer table search bar

This patch updates the table widget to allow for a placeholder
attribute and value to be set on the search input field. If not
specified, "Filter" is used as the default placeholder text. There is
also a SCSS mixin to style the placeholder text for various browsers,
which has been added to a new common _mixins.scss file.

Closes-Bug: 1433801

Change-Id: I4fdb57474b06ab17296d228becddf051f033556b