Updated openstack/openstack Project: openstack/tuskar  ab7b2199afa4dc601296e4a21a83e72030c27927
Adds tuskar-delete-roles command

The current Tuskar API doesn't cover the role resource lifecycle.
For development purposes there is no easy way to delete existing
roles that were created with tuskar-load-role and friends.

This review adds a tuskar-delete-roles utility that takes a --uuids
list of ids to delete. If any of the specified role uuids do not
exist, an unknownUUID exception is raised. If any of the specified
uuids are in use by an existing plan, a OvercloudRoleInUse exception
is raised. In case of either of the above, no roles are deleted and
the operation exits.

The --dryrun flag can be specified which still goes through the
validation above but still does not perform any deletions when
that passes.

Example invocation:

tuskar-delete-roles --config-file ~/tuskar.conf --dryrun --uuids
    c3b43c3c-e69b-4db7-b0df-3bf2bb8c04cf
    5cffe8b6-df60-4cc6-92c6-c8067f0067eb

Change-Id: Idb397fd0d6d35b143646d42b171deebc1bce6b36