Updated openstack/openstack Project: openstack/neutron  7620159c78966c6b1df8f7aac7cfa44c1aea1c76
Always fill UDP checksums in DHCP replies

In some cases the UDP checksums in packets from DHCP servers are
incorrect. This is a problem for some DHCP clients that ignore
packets with bad checksums. This patch inserts an iptables rule
to ensure DHCP servers always send packets with correct checksums.

Change-Id: I130fe0f2389bdc42eb8c858ea35dd840abecc2e7
Closes-Bug: 1244589