Updated openstack/openstack Project: openstack/horizon  e4ba59a909f58c6b5673a3364d6064b855c1eb8d
Source step - fix dynamic column headers and data

The source step of the Launch Instance Wizard contains a transfer
table which is dynamic in nature. The table's column headers and data
change based on the boot source that is selected. This patch fixes
the column headers and data mapping when a boot source is selected,
based on the latest Invision mockups (where valid).

Closes-Bug: 1432883

Change-Id: I2dd5d246471d12dd7a8c5287f1e5a6093a3c196f