Updated openstack/openstack Project: openstack/taskflow  97958ec6a0347aff7aff8654bc2c154242c66719
Use binary/encode decode helper routines in dir backend

Instead of using smaller versions of those routines just
call into the same helper routines that the rest of the
code uses.

This also allows that backend to take a configuration
option that specifies the encoding that will be used for
when translating unicode strings into binary (and vice
versa); this defaults to utf-8 to retain the existing
behavior.

Change-Id: Icfcbb22310e163753d8b34af212efc166c82634b