Updated openstack/openstack Project: openstack/python-neutronclient  c6bcc051d7ef566d5b92f29dd6bfa5a69b3371e4
Fix failures when calling list operations using Python binding

This bug is caused because of the regression caused by
I1b719bed406b83c5f2deac06e127798a91f51ad7 . The original bug was
raised because check_max_uri was not working after the introduction
of sessionclient.  The change created a regression in which the
python bindings of the neutron client was not usable and causing
traces.

The fix is to revert the change id and add a new check_max_uri to
the sessionclient.  Now uri length will be checked in sessionclient
and  httpclient as well.  please see bug for further info.

Closes-Bug: #1431449

Change-Id: Ief2352a90bb75a76e8c671d51beb0fb7a53a22f9