Updated openstack/openstack Project: openstack/neutron  7debde784328475b734cb9a8b1f085fc62bdb4ec
Validate when DVR enabled, l2_pop is also enabled

The agent should fail to start when
enable_distributed_routing = True and l2_population = False
otherwise the router won't behave as expected.

All the cross subnet traffic (between VMs of the same tenant)
in DVR is now handled locally on the compute node using the router namespace.
A Linux namespace is created for every virtual router,
on each compute node that hosts VMs that are connected to that router.

The local DVR performs the routing and replaces the source MAC address and
the destination MAC address for every packet leaving the compute node.
In order to do that correctly, the router must be populated with
all the L2 addresses of the attached networks

Change-Id: Id2b56d852eff4773a347f490b19ad9eefc86af9c
Closes-Bug: #1417633
Co-Authored-By: Romil Gupta <romilg@hp.com>