Updated openstack/openstack Project: openstack/python-neutronclient  b56c6de32850d4287a29076ab98077a5458a0ed8
Show rules in handy format in security-group-list

Previously security group rules in security-group-list was just
JSON dump of security group rules if security_group_rules field
is requested to display and it leads to long lines.

This commit defines a formater for sg rules for ListSecurityGroup.
It changes the default columns to show "security_group_rules".
I believe it is useful for users because we don't need to run
neutron security-group-rule-list to know rules.

Closes-Bug: #1153766
Change-Id: I5e1b93bea6ab1121f85dc19d2e75fffd065cd627