Updated openstack/openstack Project: openstack/trove  8eddf7e56849ff7e7be7f0c20f6c69733bfc6e57
bypass_url required in nova admin client

In the case of periodic tasks run in the task manager via the
admin client, a bypass_url is needed to be passed on client
connect in order to fix the endpoint should the token expire.
Arose as part of investigation into bug 1289101 and associated
review.

Change-Id: I15e8a40e271d5b428eebfed6e99ec47a61611795
Closes-Bug: #1431853
Co-Authored-By: Li Ma <skywalker.nick@gmail.com>
Related-Bug: #1289101