Updated openstack/openstack Project: openstack/nova  c550ad253ec9ae82815eb37740b6a32585f69b2a
Enable retry when there are multiple force hosts/nodes

Currently, when there are forced hosts/nodes in a scheduler request,
retry will be disabled. This is not right as there might be multiple
forced hosts/nodes, we should not disable retry for such case.

Closes-Bug: #1343946

DocImpact
1) If there is only one force host/node, then retry filter will not
take effect for such case.
2) If there are multiple force hosts/nodes, retry filter will help
retry on the force hosts/nodes if there are VM boot failure.

Change-Id: I6831cab70a70ebbb9609689d9132c731ff84c665