Updated openstack/openstack Project: openstack/tooz  5cf98dd27498a517706b78ce867718c8ab458f82
Use a sentinel connection pool to manage failover

When configured to use sentinel with the redis driver, allow the
redis-py client to manage the connection to the currently elected
master.

'master_for' will return a StricRedis client which is bound to a
connection pool that queries the Sentinel[s] when providing
connections from the pool.

This means that failover handling is automatic as long as the
sentinels can be reached and they have elected a new master.

Change-Id: I8fd672e664d98097944a7c984cadab5fb08dd2d6
Closes-Bug: #1434043