Updated openstack/openstack Project: openstack/sahara  c79bb88d7b9ea74edfd2dc013a8ca161e20c9fd5
Fix order of arguments in assertEqual - Part3

This part contains changes in tests of service.

Some tests used incorrect order assertEqual(observed, expected).
The correct order expected by testtools is:
    def assertEqual(self, expected, observed, message='')

It requires big changes and some tests change frequently.
So I would do it in series of small CR, and this is the last one.

Closes-Bug: #1259292

Change-Id: I20c215da18775e9a4d257462446f4adbecd26056