Updated openstack/openstack Project: openstack/nova  2fbb2c5e40fc83a211708ec2c5b8666d294d5727
Fix orphaned ports on build failure

In certain cases compute does not clean up neutron ports after
unsuccessful vm spawn.
commit 5120c4f7c2670eaa71898fe6941029bbb0081949 assumes that
deallocation is already handled in this code path
however it's not always the case (see bug report for details)
This patch adds the check for network_info is empty at the moment
failure occures. If it's empty it's better to cleanup network
to eliminate the chance of orphaned ports in neutron.

Closes-Bug: #1423845
Change-Id: I88f535193dbd35253a4444950f6b2812e1a2a407